+++ src/meta.ts
@@ -1,0 +1,5 @@
+// IMPORTANT!
+// This file has been automatically generated,
+// in order to update its content execute "pnpm run update"
+export const name = "eslint-plugin-svelte" as const
+export const version = "2.29.0" as const