+++ tests/fixtures/rules/.eslintrc.js
@@ -20,6 +20,7 @@ module.exports = {
     "one-var": "off",
     "func-style": "off",
     "no-console": "off",
+    "no-use-before-define": "off",
     "node/no-unsupported-features/es-syntax": "off",
 
     "@typescript-eslint/await-thenable": "off",