+++ tests/fixtures/rules/no-immutable-reactive-statements/valid/unknown01-input.svelte
@@ -1,0 +1,3 @@
+<script>
+  $: console.log(unknown)
+</script>