+++ tests/fixtures/rules/no-reactive-reassign/valid/member-key01-input.svelte
@@ -1,0 +1,10 @@
+<script>
+  let object = { value: 0 }
+  $: reactiveValue = { key: "value", value: object.value * 2 }
+
+  function handleClick() {
+    object[reactiveValue.key] = 42
+  }
+</script>
+
+{reactiveValue}