+++ tests/fixtures/rules/no-reactive-reassign/valid/on-dir01-input.svelte
@@ -1,0 +1,8 @@
+<script>
+  let value = 0
+  $: reactiveFn =
+    value % 2 ? () => console.log("odd") : () => console.log("even")
+</script>
+
+<button on:click={reactiveFn} />
+<input type="number" bind:value />