+++ tests/fixtures/rules/valid-each-key/valid/expression-key01-input.svelte
@@ -1,0 +1,13 @@
+<script>
+  let things = [
+    { id: 1, name: "apple" },
+    { id: 2, name: "banana" },
+    { id: 3, name: "carrot" },
+    { id: 4, name: "doughnut" },
+    { id: 5, name: "egg" },
+  ]
+</script>
+
+{#each things as thing (`thing_id=${thing.id}`)}
+  {thing.name}
+{/each}